music-player/requirements.txt
mutagen
Pillow
pygame